refactor: Renamed the useRegister file
